updated the vales in camera gimbal driver

+++ OpenCR_Core_ROS2_Fixed_Timing/camera_gimbal_driver.h
@@ -28,9 +28,9 @@ private:
 	// AX-12A (Protocol 1.0) control table
 	static constexpr float AX_PROTOCOL_VERSION = 1.0f;
 	static constexpr float XM_PROTOCOL_VERSION = 2.0f;
-	static constexpr uint8_t AX_REG_TORQUE_ENABLE = 24;     // 1 byte
-	static constexpr uint8_t AX_REG_GOAL_POSITION = 30;     // 2 bytes
-	static constexpr uint8_t AX_REG_PRESENT_POSITION = 36;  // 2 bytes
+	static constexpr uint16_t AX_REG_TORQUE_ENABLE = 24;     // 1 byte
+	static constexpr uint16_t AX_REG_GOAL_POSITION = 30;     // 2 bytes
+	static constexpr uint16_t AX_REG_PRESENT_POSITION = 36;  // 2 bytes
 
 	Dynamixel2Arduino &dxl;
 	bool inited_ = false;
